Resolve hosted deploy review blockers
The first PR pass exposed three narrow review issues: Cloud SQL port query parameters stayed in the normalized URL after being consumed, the postgres.js helper type was narrower than the wrapped API, and the Tailscale PATH isolation helper was duplicated across CLI tests. Addressing these as a separate linear commit keeps the fork PR update fast-forwardable without force-pushing.

The Storybook fetch mock also preserves extra fetch fields without referencing a canary-only missing property directly, which fixes the failed release canary build.

Constraint: Local git hook blocks force-pushes, so review fixes must land as a fast-forward commit
Constraint: Canary TypeScript does not expose fetch.preconnect on the window.fetch type
Rejected: Leave the duplicate CLI test helper | Greptile flagged the duplicated isolation logic and a shared helper is smaller overall
Confidence: high
Scope-risk: narrow
Directive: Keep Cloud SQL query parameters out of normalized postgres.js URLs once they are consumed into options
Tested: pnpm exec vitest run packages/db/src/client.test.ts cli/src/__tests__/network-bind.test.ts cli/src/__tests__/onboard.test.ts; pnpm -r typecheck; pnpm --filter @paperclipai/ui build; pnpm build
Not-tested: Full pnpm test:run after this narrow Greptile follow-up commit
