+++ cli/src/__tests__/network-bind.test.ts
@@ -1,25 +1,7 @@
-import fs from "node:fs";
-import os from "node:os";
-import path from "node:path";
 import { describe, expect, it } from "vitest";
 import { resolveRuntimeBind, validateConfiguredBindMode } from "@paperclipai/shared";
 import { buildPresetServerConfig } from "../config/server-bind.js";
-
-function withNoTailscaleOnPath<T>(run: () => T): T {
-  const originalPath = process.env.PATH;
-  const emptyBinDir = fs.mkdtempSync(path.join(os.tmpdir(), "paperclip-no-tailscale-"));
-  process.env.PATH = emptyBinDir;
-  try {
-    return run();
-  } finally {
-    if (originalPath === undefined) {
-      delete process.env.PATH;
-    } else {
-      process.env.PATH = originalPath;
-    }
-    fs.rmSync(emptyBinDir, { recursive: true, force: true });
-  }
-}
+import { withNoTailscaleOnPath } from "./test-helpers.js";
 
 describe("network bind helpers", () => {
   it("rejects non-loopback bind modes in local_trusted", () => {
@@ -67,10 +49,10 @@ describe("network bind helpers", () => {
     delete process.env.PAPERCLIP_TAILNET_BIND_HOST;
   });
 
-  it("falls back to loopback when no tailscale address is available for tailnet presets", () => {
+  it("falls back to loopback when no tailscale address is available for tailnet presets", async () => {
     delete process.env.PAPERCLIP_TAILNET_BIND_HOST;
 
-    const preset = withNoTailscaleOnPath(() =>
+    const preset = await withNoTailscaleOnPath(() =>
       buildPresetServerConfig("tailnet", {
         port: 3100,
         allowedHostnames: [],
